+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `expect_err()` method for extracting `Err` value with custom panic message
 - `expect_option()` method for converting to `Option<T>` with custom panic message on `Err`
 - Derived traits `Copy`, `PartialOrd`, `Ord`, and `Hash` for `ResultOption`
+- The `must_use` attribute to `ResultOption` enum.
 
 ## [0.2.0] - 2025-09-29